+++ dropwizard-core/src/main/java/com/yammer/dropwizard/config/DropwizardResourceConfig.java
@@ -3,6 +3,7 @@ package com.yammer.dropwizard.config;
 import com.sun.jersey.api.core.DefaultResourceConfig;
 import com.sun.jersey.api.core.ResourceConfig;
 import com.yammer.dropwizard.jersey.LoggingExceptionMapper;
+import com.yammer.dropwizard.jersey.caching.CacheControlledResourceMethodDispatchAdapter;
 import com.yammer.metrics.jersey.InstrumentedResourceMethodDispatchAdapter;
 
 public class DropwizardResourceConfig extends DefaultResourceConfig {
@@ -11,5 +12,6 @@ public class DropwizardResourceConfig extends DefaultResourceConfig {
         getFeatures().put(ResourceConfig.FEATURE_DISABLE_WADL, Boolean.TRUE);
         getSingletons().add(new LoggingExceptionMapper<Throwable>() { }); // create a subclass to pin it to Throwable
         getClasses().add(InstrumentedResourceMethodDispatchAdapter.class);
+        getClasses().add(CacheControlledResourceMethodDispatchAdapter.class);
     }
 }