+++ dropwizard-core/src/main/java/com/yammer/dropwizard/jersey/caching/CacheControlledResourceMethodDispatchAdapter.java
@@ -1,0 +1,11 @@
+package com.yammer.dropwizard.jersey.caching;
+
+import com.sun.jersey.spi.container.ResourceMethodDispatchAdapter;
+import com.sun.jersey.spi.container.ResourceMethodDispatchProvider;
+
+public class CacheControlledResourceMethodDispatchAdapter implements ResourceMethodDispatchAdapter {
+    @Override
+    public ResourceMethodDispatchProvider adapt(ResourceMethodDispatchProvider provider) {
+        return new CacheControlledResourceMethodDispatchProvider(provider);
+    }
+}